WIP: hacking on major updates to the system.
Added the `raft-essentials.md`. This is an excellent implementation
reference, and is far better than having to re-read the Raft spec
looking for details on exact semantics and nuances for implementation.

Lots of progress. RaftCore is in much better shape now. A few more core
related modules to knock out, then on to replication.

Good progress on the updates to the leader control loop.

Core loops are in place and compiling. A few items to fix.

Snapshot routine is GREATLY simplified and has a new API overall. This
should be MUCH easier to work with from the perspective of implementing
a storage engine.

A fair bit of reoganization of the code. Moved types to more logical
locations now that actix is no longer being used.

WOOT WOOT!!! The new pattern for committing client requests to the
cluster, as well as the pattern used for committing a new leader's initial
payload, and handling the cases where the new payload will commit a
previously uncommitted config change; all of these patterns are greatly
simplified and far more reliable.

The above also includes automatic stepdown of a leader when it is the
leader of a cluster from which it has been removed.

On to the last leg of the refactor. Time to update the replication
system.

[ci skip]